devided into react components
