+++ frontend/src/components/LeftSidebar/LeftSidebar.module.scss
@@ -1,0 +1,17 @@
+/* Sidebar (contacts list) with fixed width */
+.leftSidebar {
+  width: 300px;
+  background-color: var(--sidebar-bg-color);
+  border-right: 1px solid #ddd;
+  padding: var(--spacing);
+  overflow-y: auto;
+}
+
+/* Mobile mode */
+@media (max-width: 768px) {
+  .leftSidebar {
+    width: 100%;
+    height: 100%;
+    border: none;
+  }
+}
